Make all sss variables halfs

+++ crest/Assets/Crest/Crest/Shaders/Ocean.shader
@@ -296,7 +296,7 @@ Shader "Crest/Ocean"
 					const float3 uv_slice_smallerLod = WorldToUV(positionWS_XZ_before);
 
 					#if !_DEBUGDISABLESHAPETEXTURES_ON
-					float sss = 0.;
+					half sss = 0.;
 					SampleDisplacements(_LD_TexArray_AnimatedWaves, uv_slice_smallerLod, wt_smallerLod, o.worldPos, sss);
 					#endif
 
@@ -313,7 +313,7 @@ Shader "Crest/Ocean"
 					const float3 uv_slice_biggerLod = WorldToUV_BiggerLod(positionWS_XZ_before);
 
 					#if !_DEBUGDISABLESHAPETEXTURES_ON
-					float sss = 0.;
+					half sss = 0.;
 					SampleDisplacements(_LD_TexArray_AnimatedWaves, uv_slice_biggerLod, wt_biggerLod, o.worldPos, sss);
 					#endif
