+++ crest/Assets/Crest/Crest/Shaders/OceanSurfaceAlpha.shader
@@ -78,7 +78,7 @@ Shader "Crest/Ocean Surface Alpha"
 				// sample displacement textures, add results to current world pos / normal / foam
 				const float2 wxz = worldPos.xz;
 				half foam = 0.0;
-				float sss = 0.;
+				half sss = 0.;
 				SampleDisplacements(_LD_TexArray_AnimatedWaves, WorldToUV(wxz), wt_smallerLod, worldPos, sss);
 				SampleDisplacements(_LD_TexArray_AnimatedWaves, WorldToUV_BiggerLod(wxz), wt_biggerLod, worldPos, sss);
